Fix #570 (SK-D3-03): widen material_kind to u32 end-to-end
`BSLightingShaderProperty.shader_type` is `u32` at the parser layer
and `GpuMaterial.material_kind` is `u32` at the GPU layer; the
importer narrowed through `u8` in between, so the assignment

  info.material_kind = shader.shader_type as u8;

silently masked any value ≥ 256. All known Bethesda shader_type
values today are 0–20 (vanilla enum) plus engine-synthesized 100/101
(Glass / EffectShader) — safe for current data, but a future
Starfield / FO4 DLC variant in the high-byte range would have
silently routed to the wrong shader branch.

Widen three field declarations from `u8` to `u32`:

  * `crates/nif/src/import/material/mod.rs:452` — internal
    `MaterialInfo.material_kind`
  * `crates/nif/src/import/mod.rs:428` — public
    `ImportedMesh.material_kind`
  * `crates/core/src/ecs/components/material.rs:110` — ECS
    `Material.material_kind`

Drop the narrowing cast at the assignment site:

  * `crates/nif/src/import/material/walker.rs:279` —
    `info.material_kind = shader.shader_type;`

Drop the redundant widening cast at the consumer site:

  * `byroredux/src/render.rs:692` — was `m.material_kind as u32`,
    now just `m.material_kind` (already u32).

Update the `world_with_variant_material(material_kind: u32)` test
helper signature; existing call sites pass `0`/`5`/`11`/`16` which
work as either u8 or u32 literals.

Regression test:
  * `material_kind_round_trips_values_above_u8_max` — assigns 256,
    0x10001, and `u32::MAX` to `MaterialInfo.material_kind` and
    asserts each round-trips verbatim. Pre-fix the field was `u8`
    so the test wouldn't even compile; the type-system change is
    the regression guard.

SIBLING audit (per the audit's checklist) — `as u8` casts in
`material.rs` / `shader.rs`: only one was a silent narrowing
(this one). Other `as u8` casts in the area cast known-bounded
sources (alpha-test func bits, glossiness-from-shininess clamps,
flags-byte extracts) where the ≥ 256 case is unreachable by
construction. No siblings flagged.

GPU layout unaffected — `GpuMaterial.material_kind` was already
`u32` (offset 88, per `crates/renderer/src/vulkan/material.rs:110`).
This fix removes the narrowing in the middle of the pipe; the
endpoints already agreed.

1724 workspace tests pass (+1), zero new warnings.

+++ byroredux/src/render.rs
@@ -689,7 +689,7 @@ pub(crate) fn build_render_data(
                 //     wood=0.7, fabric=0.95) cleanly separates actual
                 //     transparent-refractive materials from
                 //     alpha-blend-for-edges. See follow-up to #515.
-                let base_material_kind = mat.map(|m| m.material_kind as u32).unwrap_or(0);
+                let base_material_kind = mat.map(|m| m.material_kind).unwrap_or(0);
                 // Engine-synthesized kinds (>= 100) are pre-classified
                 // upstream and must win over the heuristic Glass branch.
                 // Today: BSEffectShaderProperty meshes arrive with
@@ -1831,7 +1831,7 @@ mod variant_pack_gating_tests {
     /// dispatches; the test caller picks 0 (default) to prove the
     /// gate skips the pack, or 5/6/14 to prove the gate lets the
     /// matching variant through.
-    fn world_with_variant_material(material_kind: u8) -> World {
+    fn world_with_variant_material(material_kind: u32) -> World {
         let mut world = World::new();
 
         // Camera entity — ActiveCamera is required by build_render_data.
